requirements.txt
aioredis==1.3.1
asgiref==3.2.7
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.7.1
Automat==20.2.0
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cloudinary==1.23.0
constantly==15.1.0
cryptography==3.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.6
django-allauth==0.42.0
django-annoying==0.10.6
django-appconf==1.0.4
django-autoslug==1.9.8
django-bootstrap3==3.3.0
django-channels==0.7.0
django-ckeditor==5.9.0
django-cloudinary-storage==0.3.0
django-countries==6.1.2
django-crispy-forms==1.9.0
django-filter==2.3.0
django-gcloud-storage==0.4.0
django-heroku==0.3.1
django-hitcount==1.3.2
django-imagekit==4.0.2
django-js-asset==1.2.2
django-jsonfield==1.4.0
django-model-utils==4.0.0
django-richtextfield==1.6
django-stdimage==5.1.1
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
google-api-core==1.22.1
google-auth==1.20.1
google-cloud-core==1.4.1
google-cloud-storage==1.30.0
google-crc32c==0.1.0
google-resumable-media==0.7.1
googleapis-common-protos==1.52.0
gunicorn==20.0.4
hiredis==1.1.0
html5lib==1.1
hyperlink==19.0.0
idna==2.10
incremental==17.5.0
mysqlclient==1.4.6
oauthlib==3.1.0
pilkit==2.0
Pillow==7.1.2
progressbar2==3.51.4
protobuf==3.13.0
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
PyPDF2==1.26.0
python-decouple==3.3
python-utils==2.4.0
python3-openid==3.2.0
pytz==2020.1
reportlab==3.5.49
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
stripe==2.50.0
swapper==1.1.2.post1
txaio==20.4.1
urllib3==1.25.10
webencodings==0.5.1
whitenoise==5.1.0
xhtml2pdf==0.2.4
zope.interface==5.1.0
